gpu: repair the Vulkan vertical slice against the current encoder geometry
The Vulkan slice test did not compile. It built its draws from
EncodedPath::quadVertices and constructed its own pipeline objects, and
neither exists any more: the encoder emits a convex bounding polygon that the
vertex shader expands from vertex_index, and the pipeline classes are
constructed from the GPU runtime device rather than a wgpu device.

The quadVertices reference is a latent break, not a new one. That symbol
exists nowhere in the tree, including on the main branch, so this target has
been uncompilable since the slice landed: it is linux-only and the macOS
lanes skip it by platform, so no lane built it until this branch's linux lane
did. Repairing it here also repairs main.

The slice now builds its geometry exactly as the Metal slice does, and the
adapters both need for that - the shader IR's explicit quad, and the
expansion of compact per-band curve references into the contiguous layout the
IR consumes - move into the shared baseline header the two slices already
share their scene through, rather than being copied a second time. Pipelines
come from the device's shared instances, matching the capture tool that
produced the committed baseline.
